custom-elements/reactions/HTMLBaseElement.html

<!DOCTYPE html>
<head>
<title>Custom Elements: CEReactions on HTMLBaseElement interface</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<meta name="assert" content="href, target of HTMLBaseElement interface must have CEReactions">
<meta name="help" content="https://html.spec.whatwg.org/#the-base-element">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/custom-elements-helpers.js"></script>
</head>
<script>

testReflectAttribute('href', 'href', '/', '/', 'http://example.com/', 'http://example.com/', 'href on HTMLBaseElement', 'base', HTMLBaseElement);
testReflectAttribute('target', 'target', '_blank', '_blank', '_self', '_self', 'target on HTMLBaseElement', 'base', HTMLBaseElement);

function testReflectAttribute(jsAtName, coAtName, jsAtValue1, coAtValue1, jsAtValue2, coAtValue2, name, elementName, interfaceName) {
  test(() => {
    let element = define_build_in_custom_element([coAtName], interfaceName, elementName);
    let instance = document.createElement(elementName, { is: element.name });

    assert_array_equals(element.takeLog().types(), ['constructed']);
    document.head.appendChild(instance);
    assert_array_equals(element.takeLog().types(), ['connected']);
    instance[jsAtName] = jsAtValue1;
    let logEntries = element.takeLog();
    assert_array_equals(logEntries.types(), ['attributeChanged']);
    assert_attribute_log_entry(logEntries.last(), { name: coAtName, oldValue: null, newValue: coAtValue1, namespace: null });
}, name + ' must enqueue an attributeChanged reaction when adding a new attribute');

  test(() => {
    let element = define_build_in_custom_element([coAtName], interfaceName, elementName);
    let instance = document.createElement(elementName, { is: element.name });
    document.head.appendChild(instance);

    assert_array_equals(element.takeLog().types(), ['constructed', 'connected']);
    instance[jsAtName] = jsAtValue1;
    assert_array_equals(element.takeLog().types(), ['attributeChanged']);
    instance[jsAtName] = jsAtValue2;
    let logEntries = element.takeLog();
    assert_array_equals(logEntries.types(), ['attributeChanged']);
    assert_attribute_log_entry(logEntries.last(), { name: coAtName, oldValue: coAtValue1, newValue: coAtValue2, namespace: null });
  }, name + ' must enqueue an attributeChanged reaction when replacing an existing attribute');
}

</script>
